Use petgraph inside DefaultRoleManager to model relations
This relation model is built after the Go implementation, but uses
edges inside the graph to model relations instead of having each
role store pointers in maps.

Edges with the `EdgeVariant::Link` are relations which in Go are modeled by the
`roles` and `users` map.
Edges with the `EdgeVariant::Match` are relations which is Go are modeled by the
`matched` and `matchedBy` map.

This change brings significant speedups in benchmarks when compared against the master branch.

Co-authored-by: valkum <r.floren@heinlein-video.de>